pyproject.toml
[project]
name = "cashflow-nicegui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy-financial>=1.0.0",
    "openpyxl>=3.1.5",
    "nicegui>=2.12.1",
    "plotly>=6.5.0",
    "pandas>=2.3.3",
    "numpy-financial>=1.0.0",
    "openpyxl>=3.1.5",
    "pydantic>=2.12.5",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "xlsxwriter>=3.2.9",
]
